requirements.txt
steamship-langchain >= 0.0.20
termcolor >= 2.2.0